refactor: исправить замечания к ФЗ 11 спринта от 18.07
