+++ src/main/resources/schema.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS "users" (
+CREATE TABLE IF NOT EXISTS "user" (
                          "id" INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
                          "email" varchar,
                          "login" varchar,
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS "users" (
                          "birthday" date
 );
 
-CREATE TABLE IF NOT EXISTS "films" (
+CREATE TABLE IF NOT EXISTS "film" (
                          "id" INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
                          "name" varchar,
                          "description" varchar,
@@ -23,7 +23,8 @@ CREATE TABLE IF NOT EXISTS "genre_type" (
 
 CREATE TABLE IF NOT EXISTS "film_genre" (
                                        "film_id" integer,
-                                       "genre_id" integer
+                                       "genre_id" integer,
+                                        PRIMARY KEY ("film_id", "genre_id")
 );
 
 CREATE TABLE IF NOT EXISTS "rating_mpa" (
@@ -31,9 +32,10 @@ CREATE TABLE IF NOT EXISTS "rating_mpa" (
                                        "name" varchar
 );
 
-CREATE TABLE IF NOT EXISTS "likes" (
+CREATE TABLE IF NOT EXISTS "like" (
                          "film_id" integer,
-                         "user_id" integer
+                         "user_id" integer,
+                          PRIMARY KEY ("film_id", "user_id")
 );
 
 CREATE TABLE IF NOT EXISTS "relationship" (
@@ -43,17 +45,17 @@ CREATE TABLE IF NOT EXISTS "relationship" (
                                 "relationship_status" boolean
 );
 
-ALTER TABLE "likes" ADD FOREIGN KEY ("film_id") REFERENCES "films" ("id");
+ALTER TABLE "like" ADD FOREIGN KEY ("film_id") REFERENCES "film" ("id");
 
-ALTER TABLE "likes" ADD FOREIGN KEY ("user_id") REFERENCES "users" ("id");
+ALTER TABLE "like" ADD FOREIGN KEY ("user_id") REFERENCES "user" ("id");
 
-ALTER TABLE "film_genre" ADD FOREIGN KEY ("film_id") REFERENCES "films" ("id");
+ALTER TABLE "film_genre" ADD FOREIGN KEY ("film_id") REFERENCES "film" ("id");
 
 ALTER TABLE "film_genre" ADD FOREIGN KEY ("genre_id") REFERENCES "genre_type" ("id");
 
-ALTER TABLE "relationship" ADD FOREIGN KEY ("friend_id") REFERENCES "users" ("id");
+ALTER TABLE "relationship" ADD FOREIGN KEY ("friend_id") REFERENCES "user" ("id");
 
-ALTER TABLE "relationship" ADD FOREIGN KEY ("user_id") REFERENCES "users" ("id");
+ALTER TABLE "relationship" ADD FOREIGN KEY ("user_id") REFERENCES "user" ("id");
 
-ALTER TABLE "films" ADD FOREIGN KEY ("rating_mpa") REFERENCES "rating_mpa" ("id");
+ALTER TABLE "film" ADD FOREIGN KEY ("rating_mpa") REFERENCES "rating_mpa" ("id");
